Struct softbuffer::Surface

pub struct Surface<D, W> { /* private fields */ }

A surface for drawing to a window with software buffers.

Implementations

impl<D: HasDisplayHandle, W: HasWindowHandle> Surface<D, W>

pub fn new(context: &Context<D>, window: W) -> Result<Self, SoftBufferError>

Creates a new surface for the context for the provided window.

pub fn window(&self) -> &W

Get a reference to the underlying window handle.

pub fn resize( &mut self, width: NonZeroU32, height: NonZeroU32, ) -> Result<(), SoftBufferError>

Set the size of the buffer that will be returned by Surface::buffer_mut.

If the size of the buffer does not match the size of the window, the buffer is drawn in the upper-left corner of the window. It is recommended in most production use cases to have the buffer fill the entire window. Use your windowing library to find the size of the window.

pub fn fetch(&mut self) -> Result<Vec<u32>, SoftBufferError>

Copies the window contents into a buffer.

Platform Dependent Behavior

On X11, the window must be visible.

On macOS, Redox and Wayland, this function is unimplemented.

On Web, this will fail if the content was supplied by a different origin depending on the sites CORS rules.

pub fn buffer_mut(&mut self) -> Result<Buffer<'_, D, W>, SoftBufferError>

Return a Buffer that the next frame should be rendered into. The size must be set with Surface::resize first. The initial contents of the buffer may be zeroed, or may contain a previous frame. Call Buffer::age to determine this.

Platform Dependent Behavior

On DRM/KMS, there is no reliable and sound way to wait for the page flip to happen from within softbuffer. Therefore it is the responsibility of the user to wait for the page flip before sending another frame.